Rename persistence to data
