+++ lib/blocs/register/register_bloc.dart
@@ -1,11 +1,10 @@
+import 'package:coffeecard/data/repositories/account_repository.dart';
 import 'package:coffeecard/generated/api/coffeecard_api.swagger.models.swagger.dart';
 import 'package:coffeecard/models/account/unauthorized_error.dart';
-import 'package:coffeecard/persistence/repositories/account_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 part 'register_event.dart';
-
 part 'register_state.dart';
 
 class RegisterBloc extends Bloc<RegisterEvent, RegisterState> {
