+++ lib/data/repositories/account_repository.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 
+import 'package:coffeecard/data/api/coffee_card_api_constants.dart';
 import 'package:coffeecard/generated/api/coffeecard_api.swagger.swagger.dart';
 import 'package:coffeecard/models/account/authenticated_user.dart';
 import 'package:coffeecard/models/account/unauthorized_error.dart';
-import 'package:coffeecard/models/http/api_error.dart';
-import 'package:coffeecard/persistence/api/coffee_card_api_constants.dart';
+import 'package:coffeecard/models/api/api_error.dart';
 import 'package:crypto/crypto.dart' show sha256;
 import 'package:logger/logger.dart';
