+++ lib/data/repositories/app_config_repository.dart
@@ -1,6 +1,6 @@
+import 'package:coffeecard/data/api/coffee_card_api_constants.dart';
 import 'package:coffeecard/generated/api/coffeecard_api.swagger.swagger.dart';
-import 'package:coffeecard/models/http/api_error.dart';
-import 'package:coffeecard/persistence/api/coffee_card_api_constants.dart';
+import 'package:coffeecard/models/api/api_error.dart';
 import 'package:logger/logger.dart';
 
 class AppConfigRepository {
