+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:coffeecard/base/strings.dart';
 import 'package:coffeecard/base/style/theme.dart';
 import 'package:coffeecard/blocs/authentication/authentication_bloc.dart';
-import 'package:coffeecard/persistence/storage/secure_storage.dart';
+import 'package:coffeecard/data/storage/secure_storage.dart';
 import 'package:coffeecard/service_locator.dart';
 import 'package:coffeecard/widgets/pages/splash_page.dart';
 import 'package:coffeecard/widgets/routers/auth_router.dart';
