+++ lib/service_locator.dart
@@ -1,10 +1,10 @@
 import 'package:chopper/chopper.dart';
+import 'package:coffeecard/data/api/coffee_card_api_constants.dart';
+import 'package:coffeecard/data/api/interceptors/authentication_interceptor.dart';
+import 'package:coffeecard/data/repositories/account_repository.dart';
+import 'package:coffeecard/data/repositories/app_config_repository.dart';
+import 'package:coffeecard/data/storage/secure_storage.dart';
 import 'package:coffeecard/generated/api/coffeecard_api.swagger.swagger.dart';
-import 'package:coffeecard/persistence/api/coffee_card_api_constants.dart';
-import 'package:coffeecard/persistence/api/interceptors/authentication_interceptor.dart';
-import 'package:coffeecard/persistence/repositories/account_repository.dart';
-import 'package:coffeecard/persistence/repositories/app_config_repository.dart';
-import 'package:coffeecard/persistence/storage/secure_storage.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
